backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
groq==0.13.0
python-dotenv==1.0.1
pydantic==2.10.3
sqlmodel==0.0.21
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator==2.2.0
psycopg2-binary==2.9.10
